+++ src/MongrationDotNet.Tests/ProductMigration.cs
@@ -4,7 +4,7 @@ namespace MongrationDotNet.Tests
 {
     public class ProductMigration : CollectionMigration
     {
-        public override Version Version => new Version(1, 1, 1, 0);
+        public override Version Version => new Version(1, 1, 1, 4);
         public override string Description => "Product migration";
         public override bool MigrateArrayValues { get; } = true;
         public override string CollectionName => TestBase.CollectionName;
